+++ test/libsolidity/syntaxTests/immutable/conditionally_initialized_multiple_writes_pass.sol
@@ -1,0 +1,12 @@
+contract C {
+    uint immutable multipleWrites;
+
+    constructor() {
+        if (true) {
+            multipleWrites = 1;
+        } else {
+            multipleWrites = 2;
+        }
+    }
+}
+// ----