+++ test/libsolidity/syntaxTests/immutable/conditionally_initialized_nested_branch_write_pass.sol
@@ -1,0 +1,15 @@
+contract C {
+    uint immutable nestedBranchWrite; // good
+
+    constructor() {
+        if (true) {
+            nestedBranchWrite = 1;
+        } else {
+            if (true)
+                nestedBranchWrite = 2;
+            else
+                nestedBranchWrite = 3;
+        }
+    }
+}
+// ----