+++ test/libsolidity/syntaxTests/immutable/conditionally_initialized_nested_multiple_writes_pass.sol
@@ -1,0 +1,16 @@
+contract C {
+    uint immutable nestedMultipleWrites;
+
+    constructor() {
+        if (true) {
+            if (true) {
+                nestedMultipleWrites = 1;
+            } else {
+                nestedMultipleWrites = 1;
+            }
+        } else {
+            nestedMultipleWrites = 1;
+        }
+    }
+}
+// ----