+++ test/libsolidity/syntaxTests/immutable/conditionally_initialized_nested_read_before_write_pass.sol
@@ -1,0 +1,17 @@
+contract C {
+    uint immutable nestedReadBeforeWrite;
+
+    constructor() {
+        if (true) {
+            nestedReadBeforeWrite = 1;
+        } else {
+            if (true) {
+                nestedReadBeforeWrite = 1;
+            } else {
+                nestedReadBeforeWrite = 1;
+                nestedReadBeforeWrite + 42; // good
+            }
+        }
+    }
+}
+// ----