+++ test/libsolidity/syntaxTests/immutable/conditionally_initialized_non_assign_init_in_branch_fail.sol
@@ -1,0 +1,13 @@
+contract C {
+    uint immutable nonAssignInitInBranch;
+
+    constructor() {
+        if (true) {
+            nonAssignInitInBranch = 1;
+        } else {
+            delete nonAssignInitInBranch; // init error
+        }
+    }
+}
+// ----
+// TypeError 3969: (171-192): Immutable variables must be initialized using an assignment.