+++ test/libsolidity/syntaxTests/immutable/conditionally_initialized_non_assign_init_in_branch_pass_fail_modify.sol
@@ -1,0 +1,14 @@
+contract C {
+    uint immutable nonAssignInitInBranch;
+
+    constructor() {
+        if (true) {
+            nonAssignInitInBranch = 1;
+        } else {
+            nonAssignInitInBranch = 2;
+            delete nonAssignInitInBranch; // modify error
+        }
+    }
+}
+// ----
+// TypeError 2718: (210-231): Immutable variables cannot be modified after initialization.