+++ test/libsolidity/syntaxTests/immutable/conditionally_initialized_read_before_write_pass.sol
@@ -1,0 +1,13 @@
+contract C {
+    uint immutable readBeforeWrite;
+
+    constructor() {
+        if (true) {
+            readBeforeWrite = 1;
+        } else {
+            readBeforeWrite = 1;
+            readBeforeWrite + 42; // good
+        }
+    }
+}
+// ----