+++ test/libsolidity/syntaxTests/immutable/conditionally_initialized_single_branch_write_fail.sol
@@ -1,0 +1,10 @@
+contract C {
+    uint immutable singleBranchWrite; // error
+
+    constructor() {
+        if (true)
+            singleBranchWrite = 1;
+    }
+}
+// ----
+// TypeError 4599: (32-49): Immutable is not initialized in every branch of control flow.