+++ test/libsolidity/syntaxTests/immutable/conditionally_initialized_single_branch_write_pass.sol
@@ -1,0 +1,11 @@
+contract C {
+    uint immutable singleBranchWrite; // good
+
+    constructor() {
+        if (true)
+            singleBranchWrite = 1;
+        else
+            singleBranchWrite = 1;
+    }
+}
+// ----